+++ cmd/hidden_lake/service/internal/config/config.go
@@ -21,6 +21,7 @@ type SConfigSettings struct {
 	fMutex              sync.Mutex
 	FMessageSizeBytes   uint64 `json:"message_size_bytes" yaml:"message_size_bytes"`
 	FQueuePeriodMS      uint64 `json:"queue_period_ms" yaml:"queue_period_ms"`
+	FQueueRandPeriodMS  uint64 `json:"queue_rand_period_ms" yaml:"queue_rand_period_ms,omitempty"`
 	FKeySizeBits        uint64 `json:"key_size_bits" yaml:"key_size_bits"`
 	FWorkSizeBits       uint64 `json:"work_size_bits,omitempty" yaml:"work_size_bits,omitempty"`
 	FLimitVoidSizeBytes uint64 `json:"limit_void_size_bytes,omitempty" yaml:"limit_void_size_bytes,omitempty"`
@@ -107,6 +108,10 @@ func (p *SConfigSettings) GetKeySizeBits() uint64 {
 
 func (p *SConfigSettings) GetQueuePeriodMS() uint64 {
 	return p.FQueuePeriodMS
+}
+
+func (p *SConfigSettings) GetQueueRandPeriodMS() uint64 {
+	return p.FQueueRandPeriodMS
 }
 
 func (p *SConfigSettings) GetLimitVoidSizeBytes() uint64 {