+++ cmd/hidden_lake/service/internal/config/config_test.go
@@ -16,21 +16,22 @@ const (
 )
 
 const (
-	tcLogging       = true
-	tcNetwork       = "test_network_key"
-	tcDownloader    = "test_downloader"
-	tcUploader      = "test_uploader"
-	tcAddressTCP    = "test_address_tcp"
-	tcAddressHTTP   = "test_address_http"
-	tcAddressPPROF  = "test_address_pprof"
-	tcPubKeyAlias1  = "test_alias1"
-	tcPubKeyAlias2  = "test_alias2"
-	tcServiceName1  = "test_service1"
-	tcServiceName2  = "test_service2"
-	tcMessageSize   = (1 << 20)
-	tcWorkSize      = 22
-	tcQueuePeriod   = 1000
-	tcLimitVoidSize = (1 << 20)
+	tcLogging         = true
+	tcNetwork         = "test_network_key"
+	tcDownloader      = "test_downloader"
+	tcUploader        = "test_uploader"
+	tcAddressTCP      = "test_address_tcp"
+	tcAddressHTTP     = "test_address_http"
+	tcAddressPPROF    = "test_address_pprof"
+	tcPubKeyAlias1    = "test_alias1"
+	tcPubKeyAlias2    = "test_alias2"
+	tcServiceName1    = "test_service1"
+	tcServiceName2    = "test_service2"
+	tcMessageSize     = (1 << 20)
+	tcWorkSize        = 22
+	tcQueuePeriod     = 1000
+	tcQueueRandPeriod = 2000
+	tcLimitVoidSize   = (1 << 20)
 )
 
 var (
@@ -54,6 +55,7 @@ const (
   work_size_bits: %d
   key_size_bits: %d
   queue_period_ms: %d
+  queue_rand_period_ms: %d
   limit_void_size_bytes: %d
   network_key: %s
   f2f_disabled: true
@@ -85,6 +87,7 @@ func testNewConfigString() string {
 		tcWorkSize,
 		testutils.TcKeySize,
 		tcQueuePeriod,
+		tcQueueRandPeriod,
 		tcLimitVoidSize,
 		tcNetwork,
 		tcAddressTCP,
@@ -239,6 +242,11 @@ func TestComplexConfig(t *testing.T) {
 
 	if cfg.GetSettings().GetQueuePeriodMS() != tcQueuePeriod {
 		t.Error("settings queue period is invalid")
+		return
+	}
+
+	if cfg.GetSettings().GetQueueRandPeriodMS() != tcQueueRandPeriod {
+		t.Error("settings queue rand period is invalid")
 		return
 	}