+++ pkg/network/message/message.go
@@ -6,13 +6,13 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/hashing"
 	"github.com/number571/go-peer/pkg/crypto/keybuilder"
 	"github.com/number571/go-peer/pkg/crypto/puzzle"
+	"github.com/number571/go-peer/pkg/crypto/random"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/go-peer/pkg/payload"
 )
 
 const (
-	// third digits of PI
-	cAuthSalt = "8214808651_3282306647_0938446095_5058223172_5359408128"
+	CSaltSize = 32
 )
 
 var (
@@ -21,16 +21,19 @@ var (
 
 type sMessage struct {
 	fProof   uint64
+	fSalt    []byte
 	fHash    []byte
 	fPayload payload.IPayload
 }
 
 func NewMessage(pSett ISettings, pPld payload.IPayload, pParallel uint64) IMessage {
-	hash := getHash(pSett.GetNetworkKey(), pPld.ToBytes())
+	salt := random.NewStdPRNG().GetBytes(CSaltSize)
+	hash := getHash(pSett.GetNetworkKey(), salt, pPld.ToBytes())
 	proof := puzzle.NewPoWPuzzle(pSett.GetWorkSizeBits()).ProofBytes(hash, pParallel)
 
 	return &sMessage{
 		fProof:   proof,
+		fSalt:    salt,
 		fHash:    hash,
 		fPayload: pPld,
 	}
@@ -43,18 +46,19 @@ func LoadMessage(pSett ISettings, pData interface{}) (IMessage, error) {
 	case []byte:
 		msgBytes = x
 	case string:
-		return LoadMessage(pSett, encoding.HexDecode(x))
+		msgBytes = encoding.HexDecode(x)
 	default:
 		return nil, ErrUnknownType
 	}
 
-	if len(msgBytes) < encoding.CSizeUint64+hashing.CSHA256Size {
+	if len(msgBytes) < encoding.CSizeUint64+CSaltSize+hashing.CSHA256Size {
 		return nil, ErrInvalidHeaderSize
 	}
 
 	proofBytes := msgBytes[:encoding.CSizeUint64]
-	gotHash := msgBytes[encoding.CSizeUint64 : encoding.CSizeUint64+hashing.CSHA256Size]
-	pldBytes := msgBytes[encoding.CSizeUint64+hashing.CSHA256Size:]
+	gotSalt := msgBytes[encoding.CSizeUint64 : encoding.CSizeUint64+CSaltSize]
+	gotHash := msgBytes[encoding.CSizeUint64+CSaltSize : encoding.CSizeUint64+CSaltSize+hashing.CSHA256Size]
+	pldBytes := msgBytes[encoding.CSizeUint64+CSaltSize+hashing.CSHA256Size:]
 
 	proofArray := [encoding.CSizeUint64]byte{}
 	copy(proofArray[:], proofBytes)
@@ -65,7 +69,7 @@ func LoadMessage(pSett ISettings, pData interface{}) (IMessage, error) {
 		return nil, ErrInvalidProofOfWork
 	}
 
-	if !bytes.Equal(gotHash, getHash(pSett.GetNetworkKey(), pldBytes)) {
+	if !bytes.Equal(gotHash, getHash(pSett.GetNetworkKey(), gotSalt, pldBytes)) {
 		return nil, ErrInvalidAuthHash
 	}
 
@@ -77,6 +81,7 @@ func LoadMessage(pSett ISettings, pData interface{}) (IMessage, error) {
 
 	return &sMessage{
 		fProof:   proof,
+		fSalt:    gotSalt,
 		fHash:    gotHash,
 		fPayload: pld,
 	}, nil
@@ -84,6 +89,10 @@ func LoadMessage(pSett ISettings, pData interface{}) (IMessage, error) {
 
 func (p *sMessage) GetProof() uint64 {
 	return p.fProof
+}
+
+func (p *sMessage) GetSalt() []byte {
+	return p.fSalt
 }
 
 func (p *sMessage) GetHash() []byte {
@@ -99,6 +108,7 @@ func (p *sMessage) ToBytes() []byte {
 	return bytes.Join(
 		[][]byte{
 			proofBytes[:],
+			p.fSalt,
 			p.fHash,
 			p.fPayload.ToBytes(),
 		},
@@ -110,7 +120,7 @@ func (p *sMessage) ToString() string {
 	return encoding.HexEncode(p.ToBytes())
 }
 
-func getHash(networkKey string, pBytes []byte) []byte {
-	authKey := keybuilder.NewKeyBuilder(1, []byte(cAuthSalt)).Build(networkKey)
+func getHash(networkKey string, pAuthSalt, pBytes []byte) []byte {
+	authKey := keybuilder.NewKeyBuilder(1, []byte(pAuthSalt)).Build(networkKey)
 	return hashing.NewHMACSHA256Hasher(authKey, pBytes).ToBytes()
 }